Create 12
Company maintains employee information as employee ID, name, designation and
salary. Allow user to add, delete information of employee. Display information of
particular employee. If employee does not exist an appropriate
message is displayed. If it is, then the system displays the employee details. Use index sequential
file to maintain the data.